Add *Block.Clone method in hclwrite
The idea behind this method is creating new blocks from existing ones,
allowing to edit the cloned object in any way we like without
affecting the original block.

Our use case is from an automated process that we have that will pick
existing resoures that need to be upgraded and create new resources
with the necessary attribute changes that need to be in place at the
same time as the old ones to enable us to have a controlled
migration. Once the "old" resources are fully migrated, another
automated process will remove them, leaving the new ones.

Signed-off-by: Leandro López (inkel) <inkel.ar@gmail.com>